JavaScript Code
JavaScript code helps in the functioning of the calculator, where the operators and operands work based on their respective functional properties. For example, '+' helps in adding values, '-' for subtraction, etc.

Here, the values are extracted from the id = display, and then operators working on such operands, i.e. the values, will give the respective result.